chore: move pyproject parsing call to core.rs
